[RA] Refactors mixfile.cpp/h.
Moves template function defintions to header.
Removes dependency on functions.h.
Removes dependency on CryptRandom global.
Adds posix Finder implementation.
